Change part % to int to remove trailing decimals

+++ PokerClub/Controllers/StatsController.cs
@@ -37,8 +37,8 @@ namespace PokerClub.Controllers
                 statsView.TotalGamesPlayed = pokerTableValuations.ToList()
                     .Where(x => x.MemberId == member.Id)
                     .Select(z => z.PokerTableId).Distinct().Count();
-                statsView.ParticipationPercentage = ((double)statsView.TotalGamesPlayed / pokerTableValuations.ToList()
-                                                         .Select(z => z.PokerTableId).Distinct().Count())*100;
+                statsView.ParticipationPercentage = (int)((double)statsView.TotalGamesPlayed / pokerTableValuations.ToList()
+                                                         .Select(z => z.PokerTableId).Distinct().Count()*100);
                 statsView.TotalBuyin = pokerTableValuations.ToList()
                     .Where(x => x.MemberId == member.Id)
                     .Where(y => y.ValueCategoryId == 1)
